Perform some fixes so the previous commit builds ok.

+++ source/root_output/Collimation_root.cpp
@@ -45,7 +45,7 @@ extern "C" void CollimatorDatabaseRootWrite(int j, char* db_name_in, int db_name
     CollimationDBOutput->CollimatorDatabaseRootOutputWrite(j, db_name_in, db_name_len, db_material_in, db_material_len, db_nsig_in, db_length_in, db_rotation_in, db_offset_in);
 }
 
-extern "C" void root_FLUKA_EnergyDeposition(int id_in, int nucleons_in, double energy_in)
+extern "C" void root_FLUKA_EnergyDeposition(int id_in, int16_t nucleons_in, double energy_in)
 {
     CollimationFLUKAOutput->CollimatorFLUKARootOutputWrite(id_in, nucleons_in, energy_in);
 }
@@ -139,7 +139,7 @@ CollimationFLUKARootOutput::CollimationFLUKARootOutput()
 	CollimationFLUKATree->Branch("energy",&energy,"energy/D");
 }
 
-void CollimationFLUKARootOutput::CollimatorFLUKARootOutputWrite(int id_in, int nucleons_in, double energy_in)
+void CollimationFLUKARootOutput::CollimatorFLUKARootOutputWrite(int id_in, int16_t nucleons_in, double energy_in)
 {
 	id = id_in;
